refactor(compile): split Set/Delete emitters out of Objects.Properties.cs (#1141)
Move the whole property/index assignment emitters (EmitSetFieldsProperty,
EmitSetFieldsPropertyStrict, EmitSetProperty, EmitToLengthBoxed,
EmitSetPropertyStrict, EmitSetIndexStrict) into RuntimeEmitter.Objects.SetProperty.cs
and the delete-property emitters (EmitDeleteProperty, EmitDeletePropertyCore,
EmitDeletePropertyStrict) into RuntimeEmitter.Objects.DeleteProperty.cs.

Objects.Properties.cs shrinks 5,222 -> 3,090 lines. Pure partial-class method
moves; no IL/behaviour change. The per-receiver-branch extraction inside the
2,028-line EmitGetProperty (method-scoped shared labels/locals) is deliberately
left as a follow-up to avoid silent IL corruption.

Part of #1094.
